Make cpio output quiet & allow debug mode
We don't need block level output, so adding the `--quiet` flag to the cpio
command.

We may also want to see all commands as they're run, so added a flagged call
to `set -x`

The strip-cpio binary output is very verbose, but doesn't add much value, so
redirect stderr to /dev/null.